adding producer implementations
