+++ poc-projects/poc-kafka-producer/src/main/java/comp/KafkaProducerApp.java
@@ -1,0 +1,11 @@
+package comp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class KafkaProducerApp {
+    public static void main(String[] args) {
+        SpringApplication.run(KafkaProducerApp.class, args);
+    }
+}
