+++ poc-projects/poc-kafka-producer/src/main/java/comp/controller/EventController.java
@@ -1,0 +1,30 @@
+package comp.controller;
+
+import comp.producer.KafkaMessagePublisher;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/producer-app")
+public class EventController {
+    @Autowired
+    private KafkaMessagePublisher publisher;
+
+    @GetMapping("/publish/{message}")
+    public ResponseEntity<?> publishMessage(@PathVariable String message) {
+        try {
+            for (int i = 0; i <= 10; i++) {
+                publisher.sendMessageToTopic(message + " : " + i);
+            }
+            return ResponseEntity.ok("message published successfully ..");
+        } catch (Exception ex) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .build();
+        }
+    }
+}
