+++ poc-projects/poc-kafka-producer/src/main/java/resources/application.yml
@@ -1,0 +1,9 @@
+server:
+  port: 9191
+
+#spring:
+#  kafka:
+#    producer:
+#      bootstrap-servers: localhost:9092
+#      key-serializer: org.apache.kafka.common.serialization.StringSerializer
+#      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
